requirements.txt
# 
# !!! Requirements file generated by nmk-base plugin, don't edit !!!
#

#
# Package dependencies from nmk project files
#

build
myst-parser
nmk-base
nmk-base>=0.26.0
nmk-vscode
pytest-cov
pytest-multilog
pytest-xdist
ruff
setuptools
sphinx
sphinx-autoapi
sphinx_rtd_theme
wheel


